modified code

+++ src/components/AdminPage/stock/StockList.jsx
@@ -2,6 +2,10 @@ import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import AdminNavbar from "../AdminNavbar";
+import { Button, Col, Container, Row } from "react-bootstrap";
+import { LinkContainer } from "react-router-bootstrap";
+import Icon from "react-crud-icons";
+
 const Stock = (props) => (
   <tr>
     <td>{props.stock.name}</td>
@@ -10,17 +14,20 @@ const Stock = (props) => (
     <td>{props.stock.sector}</td>
     <td>{props.stock.exchange_id}</td>
     <td>
-      <Link className="btn btn-link" to={`/stock/edit/${props.stock._id}`}>
-        Edit
+      <Link className="btn btn-link text-decoration-none" to={`/stock/edit/${props.stock._id}`}>
+        <Icon name="edit" tooltip="Edit"   theme="light" size="tiny" />
       </Link>{" "}
       |
       <button
-        className="btn btn-link"
+         style={{"border":"none", "size":"1000px"}}
         onClick={() => {
-          props.deleteStock(props.stock._id);
+          const confirmBox = window.confirm("Do you really want to delete ");
+          if (confirmBox === true) {
+            props.deleteStock(props.stock._id);
+          }
         }}
       >
-        Delete
+         <Icon name="delete" tooltip="Delete"   theme="light" size="tiny" />
       </button>
     </td>
   </tr>
@@ -32,7 +39,7 @@ const StockList = () => {
   // This method fetches the stocks from the database.
   useEffect(() => {
     const getStocks = async () => {
-    await axios
+      await axios
         .get(`http://localhost:5000/stock/`)
         .then((res) => {
           const stocks = res.data;
@@ -75,20 +82,34 @@ const StockList = () => {
   return (
     <div>
       <AdminNavbar />
-      <h3>Stock List</h3>
-      <table className="table table-striped" style={{ marginTop: 20 }}>
-        <thead>
-          <tr>
-            <th>Name</th>
-            <th>Code</th>
-            <th>Description</th>
-            <th>Sector</th>
-            <th>Exchange</th>
-            <th>Action</th>
-          </tr>
-        </thead>
-        <tbody>{stockList()}</tbody>
-      </table>
+      <Container>
+        <Row>
+          <Col
+            lg={16}
+            md={16}
+            sm={12}
+            className="p-5 m-auto shadow-sm rounded-lg .ml-3"
+          >
+            <h3>Stock List</h3>{" "}
+            <LinkContainer to="/addstock">
+              <Button>Add Stock</Button>
+            </LinkContainer>
+            <table className="table table-striped" style={{ marginTop: 20 }}>
+              <thead>
+                <tr>
+                  <th>Name</th>
+                  <th>Code</th>
+                  <th>Description</th>
+                  <th>Sector</th>
+                  <th>Exchange</th>
+                  <th>Action</th>
+                </tr>
+              </thead>
+              <tbody>{stockList()}</tbody>
+            </table>
+          </Col>
+        </Row>
+      </Container>
     </div>
   );
 };
